# perf(Channel): linear speed position getter