**Please describe the changes this PR makes and why it should be merged:**

Similar to #9493, but for channels.

|   Runs |                                        Old |                                           New |                             Difference |
| :----: | :----------------------------------------: | :-------------------------------------------: | :------------------------------------: |
|  10000 |  83.95ms (m: 3.772μs M: 1.06ms ~: 6.902μs) |  31.84ms (m: 0.838μs M: 327.167μs ~: 1.801μs) | 37.93% (m: 22.22% M: 30.86% ~: 26.09%) |
| 100000 | 708.62ms (m: 3.702μs M: 0.46ms ~: 5.764μs) | 301.75ms (m: 0.489μs M:  38.975μs ~: 1.767μs) | 42.58% (m: 13.21% M:  8.47% ~: 30.66%) |

<details>
<summary>Labels</summary>

- `m`: minimum time per run
- `M`: maximum time per run
- `~`: mean time per run

</details>

Basically skips the `sorted` call, which is an algorithm with an O(n * log(n)) cost, a filter O(n), as well as the indexOf, which is another O(n) call. With this, not only it's not longer allocating 2 Collections nor an Array (making the only allocation be the iterator used in reduce), but the performance also goes down to a single O(n).

Also removed a repeated `indexOf` call for the target channel on every iteration, providing an extra performance boost. This code should see a performance boost between x4 and x5.

I also added `GuildForum` as a compatible type, and renamed variables to be more understandable of the actual behaviour they see.

<!--
copilot:poem
-->
### <samp>🤖 Generated by Copilot at e0c74a8</samp>

> _We're sailing on the Discord sea, with channels of all kinds_
> _We've got to sort them by their type, and keep them well aligned_
> _So heave away, me hearties, heave away with all your might_
> _We'll use the `ChannelType` enum to get the order right_

**Status and versioning classification:**

<!--
- Code changes have been tested against the Discord API, or there are no code changes
-->
- I know how to update typings and have done so, or typings don't need updating